Add URLProtectionsSpace extension
